+++ rusty_types/base.py
@@ -1,0 +1,34 @@
+from typing import Any, Dict, List
+
+
+class _Uninstantiable:
+    def __new__(self, *args, **kwargs):
+        raise TypeError("Cannot instantiate %r" % self.__class__)
+
+
+class _BaseMeta(type):
+    """ Metaclass for every type defined here.
+
+    """
+    def __init__(self, *args, **kwargs):
+        pass
+
+    def __repr__(self):
+        return "{}.{}".format(self.__module__, self.qualname)
+
+    @property
+    def qualname(self):
+        return getattr(self, "__qualname__", self.__name__)
+
+
+class _BasePositional:
+    def __init__(self, value=None):
+        self._value = value
+
+    @property
+    def value(self):
+        return self._value
+
+    @property
+    def qualname(self):
+        return getattr(self, "__qualname__", self.__name__)