+++ rusty_types/either.py
@@ -1,39 +1,4 @@
-#! /usr/bin/env python3
-
-from typing import Any, Dict, List
-
-
-class _Uninstantiable:
-    def __new__(self, *args, **kwargs):
-        raise TypeError("Cannot instantiate %r" % self.__class__)
-
-
-class _BaseMeta(type):
-    """ Metaclass for every type defined here.
-
-    """
-    def __init__(self, *args, **kwargs):
-        pass
-
-    def __repr__(self):
-        return "{}.{}".format(self.__module__, self.qualname)
-
-    @property
-    def qualname(self):
-        return getattr(self, "__qualname__", self.__name__)
-
-
-class _BasePositional:
-    def __init__(self, value=None):
-        self._value = value
-
-    @property
-    def value(self):
-        return self._value
-
-    @property
-    def qualname(self):
-        return getattr(self, "__qualname__", self.__name__)
+from rusty_types.base import _BaseMeta, _BasePositional, _Uninstantiable
 
 
 class Left(_BasePositional):
@@ -116,42 +81,3 @@ class _EitherMeta(_BaseMeta):
 class Either(_Uninstantiable, metaclass=_EitherMeta):
     __left_type__ = None
     __right_type__ = None
-
-
-class Ok(_BasePositional):
-    def is_ok(self):
-        return True
-
-    def is_err(self):
-        return False
-
-
-class Err(_BasePositional):
-    def is_ok(self):
-        return False
-
-    def is_err(self):
-        return True
-
-
-class _ResultMeta(_EitherMeta):
-    __left_class__ = Ok
-    __right_class__ = Err
-
-    def __subclasscheck__(self, cls):
-        if cls is Any:
-            return True
-
-
-class Result(Either, metaclass=_ResultMeta):
-    pass
-
-
-print(Result)
-print(isinstance(Ok(1), Result[int, str]))
-print(not isinstance(Ok("asd"), Result[int, str]))
-print(isinstance(Err("asd"), Result[int, str]))
-print(not isinstance(Err(1), Result[int, str]))
-print(not isinstance(Ok(1), Result))
-print(isinstance(Ok(), Result[None, List[Dict[str, Any]]]))
-print(isinstance(Err([{"foo": "bar"}]), Result[None, List[Dict[str, Any]]]))