+++ tests/test_either.py
@@ -1,0 +1,14 @@
+from rusty_types.either import Either, Left, Right
+
+
+def test_foo():
+    pass
+
+# print(Result)
+# print(isinstance(Ok(1), Result[int, str]))
+# print(not isinstance(Ok("asd"), Result[int, str]))
+# print(isinstance(Err("asd"), Result[int, str]))
+# print(not isinstance(Err(1), Result[int, str]))
+# print(not isinstance(Ok(1), Result))
+# print(isinstance(Ok(), Result[None, List[Dict[str, Any]]]))
+# print(isinstance(Err([{"foo": "bar"}]), Result[None, List[Dict[str, Any]]]))