+++ CHANGELOG.md
@@ -1,2 +1,19 @@
 # Change Log
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
+
+## [0.1.0] - Unreleased
+### Added
+- rusty_types/base.py
+  - _Uninstantiable - TODO
+  - _BaseMeta - TODO
+  - _BasePositional - TODO
+- rusty_types/either.py
+  - Either - A class that can contain a left or right type
+  - Left - TODO
+  - Right - TODO
+  - _EitherMeta - TODO
+- rusty_types/result.py
+  - Result - A class that can contain an ok or err type
+  - Ok - TODO
+  - Err - TODO
+  - _ResultMeta - TODO